backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
chromadb==0.4.18
sentence-transformers==2.2.2
langchain>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
pypdf==3.17.4
python-docx==1.1.0
watchdog==3.0.0
psutil==5.9.6
schedule==1.2.0
python-dotenv==1.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0 